requirements.txt
flask
flask_wtf
flask_sqlalchemy
flask_debugtoolbar
Flask-RQ2
flask-admin
flask-login
rq-dashbaord
WTForms-Components
configobj
gunicorn
pandas
xlrd
psycopg2
backports-datetime-fromisoformat
